requirements.txt
opencv-python~=4.12.0.88
numpy~=2.2.6
crccheck~=1.3.1
pillow~=11.3.0
fastapi~=0.116.1
slowapi~=0.1.9
starlette~=0.47.2
requests~=2.32.4
fpdf2~=2.8.3
webcolors~=24.11.1
scikit-learn~=1.7.1